chore: Move nft state logic into new store
