+++ src/redux/actionTypes.ts
@@ -12,10 +12,8 @@ export const UPDATE_SPAWNING_POOL_INFO = "updateSpawningPoolInfo"
 export const UPDATE_SPAWNING_POOL_USER_INFO = "updateSpawningPoolUserInfo"
 export const UPDATE_DR_FRANKENSTEIN_ZOMBIE_BALANCE = "updateDrFrankensteinZombieBalance"
 export const UPDATE_DR_FRANKENSTEIN_TOTAL_ALLOC_POINT = "updateDrFrankensteinTotalAllocPoint"
-export const UPDATE_NFT_TOTAL_SUPPLY = "updateNftTotalSupply"
 export const UPDATE_AUCTION_INFO = "updateAuctionInfo"
 export const UPDATE_AUCTION_USER_INFO = "updateAuctionUserInfo"
-export const UPDATE_NFT_USER_INFO = "updateNftUserInfo"
 export const UPDATE_BNB_BALANCE = "updateBnbBalance"
 export const UPDATE_TOMB_OVERLAY_POOL_INFO = "updateTombOverlayPoolInfo"
 export const UPDATE_TOMB_OVERLAY_USER_INFO = "updateTombOverlayUserInfo"
