+++ src/views/Graveyard/components/TransferNftModal.tsx
@@ -6,9 +6,9 @@ import { Button, Input, Modal, Text } from '@rug-zombie-libs/uikit'
 import useToast from 'hooks/useToast'
 import { useERC721 } from 'hooks/useContract'
 import InfoRow from './InfoRow'
-import { nftById } from '../../../redux/get'
 import { ToastDescriptionWithTx } from '../../../components/Toast'
 import { getAddress } from '../../../utils/addressHelpers'
+import { useGetNftById } from '../../../state/hooks'
 
 interface TransferNftModalProps {
   id: number
@@ -43,7 +43,7 @@ const TransferNftModal: React.FC<TransferNftModalProps> = ({ id, tokenId, onSucc
   const [value, setValue] = useState('')
   const [error, setError] = useState(null)
   const { account } = useWeb3React()
-  const { address, name} = nftById(id)
+  const { address, name} = useGetNftById(id)
   const contract = useERC721(getAddress(address))
   const { toastSuccess } = useToast()
