+++ src/views/Home/components/ZmbeStats.tsx
@@ -6,7 +6,7 @@ import { useTotalSupply, useBurnedBalance } from 'hooks/useTokenBalance'
 import { useTranslation } from 'contexts/Localization'
 import { getZombieAddress } from 'utils/addressHelpers'
 import CardValue from './CardValue'
-import { nftTotalSupply } from '../../../redux/get'
+import { useGetNftTotalSupply } from '../../../state/hooks'
 
 const StyledCakeStats = styled(Card)`
   margin-left: auto;
@@ -48,7 +48,7 @@ const ZmbeStats: React.FC = () => {
         </Row>
         <Row>
           <Text fontSize="14px">{t('Total NFT\'s Minted')}</Text>
-          <CardValue fontSize="14px" decimals={0} value={nftTotalSupply().toNumber()} />
+          <CardValue fontSize="14px" decimals={0} value={useGetNftTotalSupply().toNumber()} />
         </Row>
       </CardBody>
     </StyledCakeStats>
