+++ src/views/SpawningPools/components/SpawningPoolTable/components/Bottom/components/TableDetails/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import styled from 'styled-components'
 import numeral from 'numeral'
-import { bnbPriceUsd, nftById, zombiePriceUsd } from '../../../../../../../../redux/get'
-import { getBalanceAmount, getBalanceNumber, getFullDisplayBalance } from '../../../../../../../../utils/formatBalance'
+import { bnbPriceUsd, zombiePriceUsd } from '../../../../../../../../redux/get'
+import { getBalanceAmount, getFullDisplayBalance } from '../../../../../../../../utils/formatBalance'
 import { SpawningPool } from '../../../../../../../../state/types'
 import { formatDays } from '../../../../../../../../utils/timerHelpers'
-import { getPoolApr } from '../../../../../../../../utils/apr'
+import { useGetNftById } from '../../../../../../../../state/hooks'
 
 export enum SpawningPoolItemType {
   Number,
@@ -94,7 +94,7 @@ const TableDetails: React.FC<TableDetailsProps> = ({ spawningPool }) => {
     nftId,
     poolInfo: { withdrawCooldown, nftMintTime, totalAmount, minimumStake, unlockFee },
   } = spawningPool
-  const { name, path, type } = nftById(nftId)
+  const { name, path, type } = useGetNftById(nftId)
   const tvl = getBalanceAmount(totalAmount.times(zombiePriceUsd()))
 
   const unlockFeeUsd = unlockFee.times(bnbPriceUsd())
